Move View.delayShowKeyboardForView() to util file and make it more reusable
