+++ app/src/main/java/io/gnosis/safe/ui/settings/app/passcode/ChangeRepeatPasscodeFragment.kt
@@ -41,7 +41,7 @@ class ChangeRepeatPasscodeFragment : BaseViewBindingFragment<FragmentPasscodeBin
     override fun onResume() {
         super.onResume()
         binding.input.setRawInputType(InputType.TYPE_CLASS_NUMBER)
-        binding.input.delayedShowKeyboardForView()
+        binding.input.delayShowKeyboardForView()
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
